# contracts/app-api-v2.openapi.yaml
openapi: 3.1.0
info:
  title: MeiGallery App API v2
  version: 1.6.0
  description: >-
    App 1.0 的公共发现契约、Auth-1 保守账号访问、Interaction-1 喜欢/关注、Membership-1
    五级会员开发基线、Message-1 平台话题 HTTP 权威闭环与 Message-2 举报、屏蔽、话题关闭安全闭环。账号能力必须同时具备
    APP_AUTH_ENABLED、版本化文档、可阅读正文 URL 和生产人机验证配置才会由 bootstrap 启用；默认关闭，
    会员、消息和安全能力使用独立开关且生产环境还要求发布门禁。只有 available 且 executable 的消息权益可执行。
    话题始终由平台运营接收，不代表真人本人入驻或回复；实时通道、媒体消息、申诉、会员申请、支付、金币和媒体访问仍未纳入本次契约。
servers:
  - url: https://{host}
    variables:
      host:
        default: api.616618.xyz
  - url: http://127.0.0.1:8787
...
paths:
  /api/v2/app/bootstrap:
    get:
      tags: [Bootstrap]
      operationId: getAppBootstrap
      summary: 获取 App 能力与发现页配置
      responses:
        '200':
          description: 当前客户端可用能力
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            X-Contract-Version:
              $ref: '#/components/headers/ContractVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v2/me:
    get:
      tags: [Account]
      operationId: getAppMe
      summary: 获取当前观看者账号和会员摘要
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 本人账号摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v2/membership/catalog:
    get:
      tags: [Membership]
      operationId: getAppMembershipCatalog
      summary: 获取版本化五级会员目录
      description: >-
        可在未登录状态读取；acquisition 只说明联系平台的人工路径，不提供购买或申请接口。
        availability=planned 的权益只能展示“规划中”，客户端不得执行。
      responses:
        '200':
          description: 当前环境启用的会员目录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipCatalogResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v2/me/entitlements:
    get:
      tags: [Membership]
      operationId: getAppMembershipEntitlements
      summary: 获取本人当前会员与 typed entitlement 快照
      description: >-
        只解析 App 会员 grant；不会自动映射旧 Web vip/svip。权限执行必须同时检查 executable=true，
        不能根据显示名称或客户端缓存自行放行。
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 当前账号的服务端权威权益快照
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipSnapshotResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v2/conversations:
    get:
      tags: [Messaging]
      operationId: listAppConversations
      summary: 获取本人的平台话题列表
      description: 只返回当前账号自己的会话；人物资料失效时返回最小占位并转为只读。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListLimit'
        - $ref: '#/components/parameters/ListCursor'
      responses:
        '200':
          description: 平台话题列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      tags: [Messaging]
      operationId: createAppConversation
      summary: 发起由平台运营接收的新话题
      description: >-
        只允许有效会员针对当前可见的 platform_managed 人物发起；按服务端上海自然日额度原子扣减。
        disclosureVersion 必须与 bootstrap 对应的当前服务说明一致，同一账号与人物只保留一个话题。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationCreateRequest'
      responses:
        '200':
          description: 已存在的话题或幂等重放
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationCreateResponse'
        '201':
          description: 新话题已创建并写入平台接收说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v2/conversations/{conversationId}:
    get:
      tags: [Messaging]
      operationId: getAppConversation
      summary: 获取本人单个平台话题摘要
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: 平台话题摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/conversations/{conversationId}/messages:
    get:
      tags: [Messaging]
      operationId: listAppConversationMessages
      summary: 按序号增量拉取本人话题消息
      description: Message-1 使用 HTTP 拉取，不承诺实时到达；消息按 sequence 严格升序。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - name: afterSequence
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: 增量消息页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationMessagesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Messaging]
      operationId: sendAppConversationMessage
      summary: 在本人有效话题中发送文本消息
      description: 只接受 text；表情作为普通 Unicode 文本，不支持图片、语音、视频、礼物或付款。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationMessageSendRequest'
      responses:
        '200':
          description: 幂等重放的原消息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationMessageSendResponse'
        '201':
          description: 文本消息已接收
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationMessageSendResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v2/conversations/{conversationId}/read:
    post:
      tags: [Messaging]
      operationId: markAppConversationRead
      summary: 更新本人话题已读序号
      description: 只允许单调前进，不能超过服务端当前 lastSequence。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationReadRequest'
      responses:
        '200':
          description: 服务端确认的已读序号
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationReadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/conversations/{conversationId}/close:
    post:
      tags: [Safety, Messaging]
      operationId: closeAppConversation
      summary: 由观看者永久关闭本人的平台话题
      description: >-
        关闭后双方只能查看历史且不能重新打开；服务端追加系统消息并释放运营分配。
        重试必须复用同一个 Idempotency-Key。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 已关闭话题或幂等重放结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationCloseResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v2/me/devices:
    get:
      tags: [Account]
      operationId: listAppDevices
      summary: 获取本人设备列表
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 最多返回最近 50 台本人设备
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v2/me/devices/{deviceId}:
    delete:
      tags: [Account]
      operationId: revokeAppDevice
      summary: 远程退出本人其他设备
      description: 幂等撤销目标设备的全部活跃 App 会话；当前设备必须使用 logout。
      security:
        - BearerAuth: []
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            pattern: '^dev_[A-Za-z0-9_-]+$'
      responses:
        '200':
          description: 设备已退出或此前已经退出
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v2/me/likes:
    get:
      tags: [Interaction]
      operationId: listAppViewerLikes
      summary: 获取本人的喜欢列表
      description: 资料已失效时只返回最小不可用占位，不返回历史公开内容。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListLimit'
        - $ref: '#/components/parameters/ListCursor'
      responses:
        '200':
          description: 本人的喜欢关系
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ViewerInteractionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v2/me/follows:
    get:
      tags: [Interaction]
      operationId: listAppViewerFollows
      summary: 获取本人的已关注列表
      description: Interaction-1 只提供关系列表；公开内容更新流与通知尚未启用。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListLimit'
        - $ref: '#/components/parameters/ListCursor'
      responses:
        '200':
          description: 本人的关注关系
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ViewerInteractionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v2/discovery/regions:
    get:
      tags: [Discovery]
      operationId: listDiscoveryRegions
      summary: 获取当前存在公开人物的地区目录
      responses:
        '200':
          description: 地区目录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  headers:
    RequestId:
      description: 服务端生成的请求追踪 ID。
      schema:
        type: string
    ContractVersion:
      description: 当前响应契约版本。
      schema:
        type: string
        const: 1.6.0
...
  responses:
...
    Unauthorized:
      description: 访问凭证缺失、无效、过期或已撤销
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: 服务暂时不可用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    MeResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/MeSummary'
        meta:
          $ref: '#/components/schemas/Meta'
...
    BootstrapResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/Bootstrap'
        meta:
          $ref: '#/components/schemas/Meta'
...